modify: rename RayObjectImpl.Serializer to RayObjectSerializer and move to outer
