+++ java/runtime/src/main/java/org/ray/runtime/util/Serializer.java
@@ -4,6 +4,7 @@ import org.nustaq.serialization.FSTConfiguration;
 import org.ray.runtime.actor.NativeRayActor;
 import org.ray.runtime.actor.NativeRayActorSerializer;
 import org.ray.runtime.object.RayObjectImpl;
+import org.ray.runtime.object.RayObjectSerializer;
 
 /**
  * Java object serialization TODO: use others (e.g. Arrow) for higher performance
@@ -13,7 +14,7 @@ public class Serializer {
   private static final ThreadLocal<FSTConfiguration> conf = ThreadLocal.withInitial(() -> {
     FSTConfiguration conf = FSTConfiguration.createDefaultConfiguration();
     conf.registerSerializer(NativeRayActor.class, new NativeRayActorSerializer(), true);
-    conf.registerSerializer(RayObjectImpl.class, new RayObjectImpl.Serializer(), true);
+    conf.registerSerializer(RayObjectImpl.class, new RayObjectSerializer(), true);
     return conf;
   });
